fixed8: NewFixed864 -> NewFixed8

+++ common/utils/fixed8_test.go
@@ -26,9 +26,9 @@ func TestFixed8FromInt64(t *testing.T) {
 }
 
 func TestNewFixed8(t *testing.T) {
-	values := []int{9000, 1e8, 5, 10945}
+	values := []int64{9000, 1e8, 5, 10945}
 	for _, val := range values {
-		assert.Equal(t, Fixed8(val*decimals), NewFixed8(val))
+		assert.Equal(t, Fixed8(val*int64(decimals)), NewFixed8(val))
 		assert.Equal(t, int64(val), NewFixed8(val).Value())
 	}
 }
